// web/admin/src/types/api/models/MessagesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Marche Online
 * マルシェ管理者用API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { MessagesResponseMessagesInner } from './MessagesResponseMessagesInner';
import {
    MessagesResponseMessagesInnerFromJSON,
    MessagesResponseMessagesInnerFromJSONTyped,
    MessagesResponseMessagesInnerToJSON,
} from './MessagesResponseMessagesInner';

/**
 * 
 * @export
 * @interface MessagesResponse
 */
export interface MessagesResponse {
    /**
     * メッセージ一覧
     * @type {Array<MessagesResponseMessagesInner>}
     * @memberof MessagesResponse
     */
    messages: Array<MessagesResponseMessagesInner>;
    /**
     * 合計数
     * @type {number}
     * @memberof MessagesResponse
     */
    total: number;
}

/**
 * Check if a given object implements the MessagesResponse interface.
 */
export function instanceOfMessagesResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "messages" in value;
    isInstance = isInstance && "total" in value;

    return isInstance;
}

export function MessagesResponseFromJSON(json: any): MessagesResponse {
    return MessagesResponseFromJSONTyped(json, false);
}

export function MessagesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessagesResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'messages': ((json['messages'] as Array<any>).map(MessagesResponseMessagesInnerFromJSON)),
        'total': json['total'],
    };
}

export function MessagesResponseToJSON(value?: MessagesResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'messages': ((value.messages as Array<any>).map(MessagesResponseMessagesInnerToJSON)),
        'total': value.total,
    };
}